Address  reviewer comments
Environment is now the union of a ASCIIString and Symbol
Look in deps.jl for default_dir in build.jl to avoid overriding when Pkg.build is called

+++ deps/build.jl
@@ -1,4 +1,15 @@
-default_dir = get(ENV, "CONDA_JL_HOME", abspath(joinpath(dirname(@__FILE__), "usr")))
-open("deps.jl", "w") do f
-    println(f, "default_dir=\"$(escape_string(default_dir))\"")
+if haskey(ENV, "CONDA_JL_HOME")
+    default_dir = ENV["CONDA_JL_HOME"]
+elseif isfile("deps.jl")
+    include("deps.jl")
+else
+    default_dir = abspath(dirname(@__FILE__), "usr")
+end
+
+deps = "default_dir=\"$(escape_string(default_dir))\""
+
+if !isfile("deps.jl") || readchomp("deps.jl") != deps
+    open("deps.jl", "w") do f
+        println(f, deps)
+    end
 end
